requirements.txt
Flask==2.3.2
firebase-admin==6.1.0
requests==2.31.0
python-dotenv==1.0.0
